IN-9 refactor: change structure components Tabs
