+++ src/shared/ui/Tabs/index.ts
@@ -1,4 +1,2 @@
-export { Tabs } from "./Tabs";
-export { TabsContent } from "./TabsContent";
-export { TabsList } from "./TabsList";
-export { TabsTrigger } from "./TabsTrigger";
+export { Tabs, TabsContent } from "./Tabs";
+export { TabsList, TabsTrigger } from "./TabsListItem";
